requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
pyyaml>=6.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0

